transaction-kafka-producer/requirements.txt
# Kafka
confluent-kafka==2.3.0
kafka-python==2.0.2

# PostgreSQL
psycopg2-binary==2.9.9

# HTTP API
requests==2.31.0

# Environment variables
python-dotenv==1.0.1
